Took RoutePathGroup out from RoutePathListTab
