+++ src/components/sidebar/routeListView/RoutePathListTab.tsx
@@ -1,20 +1,10 @@
-import classnames from 'classnames';
 import _ from 'lodash';
 import { inject, observer } from 'mobx-react';
 import Moment from 'moment';
 import React from 'react';
-import { FaTrashAlt } from 'react-icons/fa';
-import { FiInfo } from 'react-icons/fi';
-import { Button } from '~/components/controls';
-import InputContainer from '~/components/controls/InputContainer';
 import SavePrompt, { ISaveModel } from '~/components/overlays/SavePrompt';
 import SaveButton from '~/components/shared/SaveButton';
-import Loader from '~/components/shared/loader/Loader';
-import TransitType from '~/enums/transitType';
 import { IRoutePath } from '~/models';
-import navigator from '~/routing/navigator';
-import routeBuilder from '~/routing/routeBuilder';
-import subSites from '~/routing/subSites';
 import RoutePathMassEditService from '~/services/routePathMassEditService';
 import RoutePathService from '~/services/routePathService';
 import { AlertStore } from '~/stores/alertStore';
@@ -24,9 +14,7 @@ import { LoginStore } from '~/stores/loginStore';
 import { MapStore } from '~/stores/mapStore';
 import { RouteListStore } from '~/stores/routeListStore';
 import { RoutePathMassEditStore } from '~/stores/routePathMassEditStore';
-import { UserStore } from '~/stores/userStore';
-import { toDateString } from '~/utils/dateUtils';
-import ToggleSwitch from '../../controls/ToggleSwitch';
+import RoutePathGroup from './RoutePathGroup';
 import * as s from './routePathListTab.scss';
 
 interface IRoutePathStopNames {
@@ -42,7 +30,6 @@ interface IRoutePathListTabProps {
     routeListStore?: RouteListStore;
     mapStore?: MapStore;
     confirmStore?: ConfirmStore;
-    userStore?: UserStore;
     loginStore?: LoginStore;
     alertStore?: AlertStore;
     errorStore?: ErrorStore;
@@ -62,7 +49,6 @@ const ROUTE_PATH_GROUP_SHOW_LIMIT = 3;
     'routeListStore',
     'mapStore',
     'confirmStore',
-    'userStore',
     'loginStore',
     'alertStore',
     'errorStore',
@@ -203,207 +189,6 @@ class RoutePathListTab extends React.Component<IRoutePathListTabProps, IRoutePat
         }
     };
 
-    // TODO: move into GroupedRoutePaths.tsx?
-    private renderGroupedRoutePaths = (groupedRoutePaths: IRoutePath[][]) => {
-        return groupedRoutePaths.map((routePaths: IRoutePath[], index) => {
-            const hasDirection1 = Boolean(routePaths.find((rp) => rp.direction === '1'));
-            const hasDirection2 = Boolean(routePaths.find((rp) => rp.direction === '2'));
-            // Group above the current group
-            const nextGroup: IRoutePath[] | null = _findNextGroup(
-                groupedRoutePaths,
-                index,
-                hasDirection1,
-                hasDirection2
-            );
-            // Group below the current group
-            const prevGroup: IRoutePath[] | null = _findPrevGroup(
-                groupedRoutePaths,
-                index,
-                hasDirection1,
-                hasDirection2
-            );
-
-            let minStartDate = undefined;
-            let maxEndDate = undefined;
-            if (prevGroup) {
-                minStartDate = _.cloneDeep(prevGroup[0].endTime);
-                minStartDate.setDate(minStartDate.getDate() + 1);
-            }
-            if (nextGroup) {
-                maxEndDate = _.cloneDeep(nextGroup[0].startTime);
-                maxEndDate.setDate(maxEndDate.getDate() - 1);
-            }
-            const isEditing = this.props.isEditing;
-            const first = routePaths[0];
-            const header = `${toDateString(first.startTime)} - ${toDateString(first.endTime)}`;
-
-            let validationResult;
-            if (isEditing) {
-                validationResult = this.props.routePathMassEditStore!.massEditRoutePaths?.find(
-                    (m) => m.routePath.internalId === first.internalId
-                )?.validationResult;
-            }
-            return (
-                <div
-                    key={header}
-                    className={classnames(s.groupedRoutes, index % 2 ? s.shadow : undefined)}
-                >
-                    <div className={s.groupedRoutesDates}>
-                        {isEditing ? (
-                            <>
-                                <InputContainer
-                                    label=''
-                                    disabled={!this.props.isEditing}
-                                    type='date'
-                                    value={first.startTime}
-                                    onChange={this.updateStartDates(routePaths)}
-                                    validationResult={validationResult}
-                                    minStartDate={minStartDate}
-                                    maxEndDate={maxEndDate}
-                                />
-                                <InputContainer
-                                    label=''
-                                    disabled={!this.props.isEditing}
-                                    type='date'
-                                    value={first.endTime}
-                                    onChange={this.updateEndDates(routePaths)}
-                                    minStartDate={minStartDate}
-                                    maxEndDate={maxEndDate}
-                                />
-                            </>
-                        ) : (
-                            <div>{header}</div>
-                        )}
-                    </div>
-                    <div className={s.groupedRoutesContent}>
-                        {this.renderRoutePathList(routePaths)}
-                    </div>
-                </div>
-            );
-        });
-    };
-
-    private updateStartDates = (routePaths: IRoutePath[]) => (value: Date) => {
-        routePaths.forEach((rp) => {
-            this.props.routePathMassEditStore!.updateRoutePathStartDate(rp.internalId, value);
-        });
-    };
-
-    private updateEndDates = (routePaths: IRoutePath[]) => (value: Date) => {
-        routePaths.forEach((rp) => {
-            this.props.routePathMassEditStore!.updateRoutePathEndDate(rp.internalId, value);
-        });
-    };
-
-    private renderRoutePathList = (routePaths: IRoutePath[]) => {
-        return routePaths.map((routePath: IRoutePath) => {
-            const toggleRoutePathVisibility = () => {
-                this.props.routeListStore!.toggleRoutePathVisibility(routePath.internalId);
-            };
-
-            const openRoutePathView = () => {
-                const routePathViewLink = routeBuilder
-                    .to(subSites.routePath)
-                    .toTarget(
-                        ':id',
-                        [
-                            routePath.routeId,
-                            Moment(routePath.startTime).format('YYYY-MM-DDTHH:mm:ss'),
-                            routePath.direction,
-                        ].join(',')
-                    )
-                    .toLink();
-                navigator.goTo({ link: routePathViewLink });
-            };
-
-            const shouldHighlightRoutePath = this.isCurrentTimeWithinRoutePathTimeSpan(routePath);
-            const stopNames = this.state.stopNameMap.get(routePath.internalId);
-            const isLoading = !stopNames && this.state.areStopNamesLoading;
-            const stopOriginFi = stopNames?.firstStopName ? stopNames.firstStopName : '-';
-            const stopDestinationFi = stopNames?.lastStopName ? stopNames?.lastStopName : '-';
-            const stopDestinations = `${stopOriginFi} - ${stopDestinationFi}`;
-            const routePathDestinations = `${routePath.originFi} - ${routePath.destinationFi}`;
-            const isEditing = this.props.isEditing;
-            const massEditRp = this.props.routePathMassEditStore!.massEditRoutePaths?.find(
-                (m) => m.routePath.internalId === routePath.internalId
-            )!;
-            return (
-                <div
-                    className={classnames(
-                        s.routePathContainer,
-                        isEditing && massEditRp.isNew ? s.highlighAsNew : undefined
-                    )}
-                    key={routePath.internalId}
-                >
-                    <div
-                        className={
-                            shouldHighlightRoutePath
-                                ? classnames(s.routePathInfo, s.highlight)
-                                : s.routePathInfo
-                        }
-                    >
-                        <div className={s.routePathDirection}>{routePath.direction}</div>
-                        <div>
-                            {isLoading ? (
-                                <Loader
-                                    containerClassName={s.stopNameLoader}
-                                    size='tiny'
-                                    hasNoMargin={true}
-                                />
-                            ) : (
-                                <>
-                                    <div className={s.destinations1}>
-                                        {this.props.userStore!.userTransitType === TransitType.BUS
-                                            ? routePathDestinations
-                                            : stopDestinations}
-                                    </div>
-                                    <div className={s.destinations2}>
-                                        {this.props.userStore!.userTransitType === TransitType.BUS
-                                            ? stopDestinations
-                                            : routePathDestinations}
-                                    </div>
-                                </>
-                            )}
-                        </div>
-                    </div>
-                    <div className={s.routePathControls}>
-                        {isEditing && massEditRp.isNew && (
-                            <Button
-                                className={s.removeNewRoutePathButton}
-                                hasReverseColor={true}
-                                onClick={this.removeNewRoutePath(routePath.internalId)}
-                            >
-                                <FaTrashAlt />
-                            </Button>
-                        )}
-                        <Button
-                            className={s.openRoutePathViewButton}
-                            hasReverseColor={true}
-                            onClick={openRoutePathView}
-                            data-cy='openRoutePathViewButton'
-                        >
-                            <FiInfo />
-                        </Button>
-                        <ToggleSwitch
-                            onClick={toggleRoutePathVisibility}
-                            value={routePath.visible}
-                            color={routePath.visible ? routePath.color! : '#898989'}
-                        />
-                    </div>
-                </div>
-            );
-        });
-    };
-
-    private removeNewRoutePath = (id: string) => () => {
-        this.props.routePathMassEditStore!.removeRoutePath(id);
-        // Remove routePath also from current props because routePaths in state doesn't get updated otherwise
-        const routePaths = _.cloneDeep(this.props.routePaths);
-        const removeIndex = routePaths.findIndex((rp) => rp.internalId === id);
-        routePaths.splice(removeIndex, 1);
-        this.updateGroupedRoutePathsToDisplay(routePaths);
-    };
-
     private isCurrentTimeWithinRoutePathTimeSpan = (routePath: IRoutePath) => {
         return (
             Moment(routePath.startTime).isBefore(Moment()) &&
@@ -447,6 +232,15 @@ class RoutePathListTab extends React.Component<IRoutePathListTabProps, IRoutePat
         this.props.routeListStore!.setRouteIdToEdit(null);
     };
 
+    private removeNewRoutePath = (id: string) => () => {
+        this.props.routePathMassEditStore!.removeRoutePath(id);
+        // Remove routePath also from current props because routePaths in state doesn't get updated otherwise
+        const routePaths = _.cloneDeep(this.props.routePaths);
+        const removeIndex = routePaths.findIndex((rp) => rp.internalId === id);
+        routePaths.splice(removeIndex, 1);
+        this.updateGroupedRoutePathsToDisplay(routePaths);
+    };
+
     render() {
         const { routePaths, isEditing } = this.props;
         if (routePaths.length === 0) {
@@ -463,7 +257,38 @@ class RoutePathListTab extends React.Component<IRoutePathListTabProps, IRoutePat
             !this.props.routePathMassEditStore!.isFormValid;
         return (
             <div className={s.routePathListTab}>
-                {this.renderGroupedRoutePaths(groupedRoutePathsToDisplay)}
+                {groupedRoutePathsToDisplay.map((routePaths: IRoutePath[], index) => {
+                    const hasDirection1 = Boolean(routePaths.find((rp) => rp.direction === '1'));
+                    const hasDirection2 = Boolean(routePaths.find((rp) => rp.direction === '2'));
+                    // Group above the current group
+                    const nextGroup: IRoutePath[] | null = _findNextGroup(
+                        groupedRoutePathsToDisplay,
+                        index,
+                        hasDirection1,
+                        hasDirection2
+                    );
+                    // Group below the current group
+                    const prevGroup: IRoutePath[] | null = _findPrevGroup(
+                        groupedRoutePathsToDisplay,
+                        index,
+                        hasDirection1,
+                        hasDirection2
+                    );
+
+                    return (
+                        <RoutePathGroup
+                            key={`routePathGroup-${index}`}
+                            routePaths={routePaths}
+                            nextGroup={nextGroup}
+                            prevGroup={prevGroup}
+                            isEditing={isEditing}
+                            stopNameMap={this.state.stopNameMap}
+                            areStopNamesLoading={this.state.areStopNamesLoading}
+                            index={index}
+                            removeNewRoutePath={this.removeNewRoutePath}
+                        />
+                    );
+                })}
                 {!isEditing && allGroupedRoutePaths.length > ROUTE_PATH_GROUP_SHOW_LIMIT && (
                     <div
                         className={s.toggleAllRoutePathsVisibleButton}
@@ -542,3 +367,5 @@ const _hasGroupRoutePathWithDirection = (
 };
 
 export default RoutePathListTab;
+
+export { IRoutePathStopNames };
